requirements.txt
fastapi
uvicorn
requests
httpx
beautifulsoup4
pydantic
trafilatura
google-genai
streamlit
